Isolate direction into it's own module
