--- projects/storefrontlib/src/layout/direction/direction.provider.ts
@@ -1,14 +1,0 @@
-import { APP_INITIALIZER, Provider } from '@angular/core';
-import { DirectionService } from './direction.service';
-
-export function initHtmlDirAttribute(directionService: DirectionService) {
-  const result = () => directionService.initialize();
-  return result;
-}
-
-export const HtmlDirProvider: Provider = {
-  provide: APP_INITIALIZER,
-  multi: true,
-  useFactory: initHtmlDirAttribute,
-  deps: [DirectionService],
-};
